pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "temporal-parseable"
dynamic = ["version"]
description = "Temporal plugin for exporting OpenTelemetry data to Parseable"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Parseable", email = "hi@parseable.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "temporalio[opentelemetry]>=1.7.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[tool.hatch.version]
path = "src/temporal_parseable/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/temporal_parseable"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
